Backport tournament domain package with 6-team RR support
