+++ jupr_app/domain/tournaments/__init__.py
@@ -1,0 +1,24 @@
+from .bracket_builder import (
+    SUPPORTED_TEAM_COUNTS,
+    build_playoff_games,
+    build_round_robin_games,
+    compute_podium_from_playoffs,
+    compute_podium_from_rr,
+    compute_round_robin_standings,
+    finalize_game,
+    resolve_playoff_dependencies,
+)
+from .sync import build_podium_payload, validate_podium_placements
+
+__all__ = [
+    "SUPPORTED_TEAM_COUNTS",
+    "build_playoff_games",
+    "build_round_robin_games",
+    "build_podium_payload",
+    "compute_podium_from_playoffs",
+    "compute_podium_from_rr",
+    "compute_round_robin_standings",
+    "finalize_game",
+    "resolve_playoff_dependencies",
+    "validate_podium_placements",
+]
