+++ jupr_app/ui/assets/tournaments/rr-6.csv
@@ -1,0 +1,16 @@
+round,slot,teamA,teamB
+1,1,1,6
+1,2,2,5
+1,3,3,4
+2,1,1,5
+2,2,6,4
+2,3,2,3
+3,1,1,4
+3,2,5,3
+3,3,6,2
+4,1,1,3
+4,2,4,2
+4,3,5,6
+5,1,1,2
+5,2,3,6
+5,3,4,5
